+++ client.go
@@ -3,10 +3,22 @@ package http
 // Client is a simplified HTTP interface that ensures that a struct is transported to a remote endpoint
 // properly encoded, and the response is decoded into the response struct.
 type Client interface {
+	Request(
+		Method string,
+		path string,
+		requestBody interface{},
+		responseBody interface{},
+	) (statusCode int, err error)
+
+	// Get queries the configured endpoint with the path providing the response in the responseBody structure. It
+	// returns the HTTP status code and any potential errors.
+	Get(
+		path string,
+		responseBody interface{},
+	) (statusCode int, err error)
+
 	// Post queries the configured endpoint with the path, sending the requestBody and providing the
 	// response in the responseBody structure. It returns the HTTP status code and any potential errors.
-	//
-	// The returned error is always one of ClientError
 	Post(
 		path string,
 		requestBody interface{},